scafflod proejct structure to follow package format
