+++ .gitignore
@@ -1,2 +1,10 @@
-.ipynb_checkpoints/
+# Python-generated files
+__pycache__/
+*.py[oc]
+build/
+dist/
+wheels/
+*.egg-info
 
+# Virtual environments
+.venv
